+++ naga/tests/out/spv/wgsl-workgroup-struct-layout.spvasm.glsl
@@ -1,0 +1,31 @@
+#version 460
+layout(local_size_x = 64, local_size_y = 1, local_size_z = 1) in;
+
+struct _8
+{
+    vec4 _m0[64];
+    uint _m1;
+};
+
+shared _8 _9;
+
+void main()
+{
+    if (gl_LocalInvocationIndex == 0u)
+    {
+        _9 = _8(vec4[](vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0), vec4(0.0)), 0u);
+    }
+    barrier();
+    _9._m0[gl_LocalInvocationIndex] = vec4(float(gl_LocalInvocationIndex));
+    barrier();
+    if (gl_LocalInvocationIndex == 0u)
+    {
+        _9._m1 = 64u;
+        return;
+    }
+    else
+    {
+        return;
+    }
+}
+